LPS-73684 Move core tests into new module
