+++ modules/apps/collaboration/repository/repository-test/src/testIntegration/java/com/liferay/repository/capabilities/test/TemporaryFileEntriesCapabilityTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.repository.capabilities;
+package com.liferay.repository.capabilities.test;
 
 import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.document.library.kernel.model.DLFolderConstants;
